Camelot.ViewModels.Windows restructuring
